Moved transitions to their primitive test file
